Seção 9 - Dtos e DataAnnotations

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -5,10 +5,10 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using ProEventos.Persistence;
-using ProEventos.Domain;
 using ProEventos.Persistence.Contexto;
 using ProEventos.Application.Contratos;
 using Microsoft.AspNetCore.Http;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -32,7 +32,7 @@ namespace ProEventos.API.Controllers
           try
           {
                var eventos = await _eventoService.GetAllEventosAsync(true);
-               if (eventos == null) return NotFound("Evento não encontrado");
+               if (eventos == null) return NoContent();
 
                return Ok(eventos);
           }
@@ -50,7 +50,7 @@ namespace ProEventos.API.Controllers
          try 
           {
                var evento = await _eventoService.GetEventoByIdAsync(id, true);
-               if (evento == null) return NotFound("Evento por id não encontrado");
+               if (evento == null) return NoContent();
 
                return Ok(evento);
           }
@@ -67,7 +67,7 @@ namespace ProEventos.API.Controllers
          try
           {
                var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-               if (evento == null) return NotFound("Evento por temas não encontrados");
+               if (evento == null) return NoContent();
 
                return Ok(evento);
           }
@@ -80,12 +80,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-       public async Task<IActionResult> Post(Evento model)
+       public async Task<IActionResult> Post(EventoDto model)
         {
          try
           {
                var evento = await _eventoService.AddEventos(model);
-               if (evento == null) return BadRequest("Erro ao tentar adicionar evento");
+               if (evento == null) return NoContent();
 
                return Ok(evento);
           }
@@ -98,12 +98,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
          try
           {
                var evento = await _eventoService.UpdateEvento(id, model);
-               if (evento == null) return NotFound("Evento atualizado não encontrados");
+               if (evento == null) return NoContent();
 
                return Ok(evento);
           }
@@ -120,9 +120,14 @@ namespace ProEventos.API.Controllers
         {
          try
           {
+              var evento = await _eventoService.GetEventoByIdAsync(id, true);
+               if (evento == null) return NoContent();
+
+
               return await _eventoService.DeleteEvento(id) ?
                 Ok("Deletado") : 
-                BadRequest("Evento não Deletado");
+                throw new Exception("Ocorreu um problema não específico ao deletar Evento");
+                
           }
           catch (Exception ex ) 
           {
